Se reestructura el proyecto para hacerlo más intuitivo y amigable, se contempla que una página puede ser un módulo, un componente o ambos.
