+++ src/app/app-routing.module.ts
@@ -3,9 +3,9 @@ import { Routes, RouterModule } from '@angular/router';
 import { AuthGuard } from './guards/auth.guard';
 
 const routes: Routes = [
-  { path: '', canActivate: [AuthGuard], loadChildren: './modules/main/main.module#MainModule' },
-  { path: 'admin', canActivate: [AuthGuard], loadChildren: './modules/admin/admin.module#AdminModule' },
-  { path: 'login', loadChildren: './modules/login/login.module#LoginModule' },
+  { path: '', canActivate: [AuthGuard], loadChildren: './pages/main/main.module#MainModule' },
+  { path: 'admin', canActivate: [AuthGuard], loadChildren: './pages/admin/admin.module#AdminModule' },
+  { path: 'login', loadChildren: './pages/login/login.module#LoginModule' },
   { path: '**', redirectTo: '', pathMatch: 'full'},
 ];
