--- src/app/modules/admin/components/pages/dashboard/dashboard.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-admin-dashboard',
-  templateUrl: './dashboard.component.html',
-  styles: []
-})
-export class DashboardComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
