+++ src/app/pages/login/login-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LoginComponent } from './components/login/login.component';
+import { LoginComponent } from './login.component';
 
 const routes: Routes = [
   { path: '', component: LoginComponent }
