+++ src/app/pages/main/main-footer/main-footer.component.ts
