+++ src/app/pages/main/main-header/main-header.component.ts
