+++ src/app/pages/main/main.component.ts
