+++ src/app/pages/main/pages/ejemplo/ejemplo.component.ts
