+++ src/app/pages/main/pages/inicio/inicio.component.html
@@ -1,3 +1,3 @@
 <div style="border:1px;border-style: solid; padding:10px;">
-    <p>dashboard component works!</p>
+    <p>inicio component works!</p>
 </div>
