+++ src/app/pages/main/pages/inicio/inicio.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-main-inicio',
+  templateUrl: './inicio.component.html',
+  styles: []
+})
+export class InicioComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
